SameSite Cookie Changes in February 2020
We can't set the value to lax here, because of the new changes to the way cookies are handled in Chrome. If same_site is set to 'lax' requests don't work in Chrome, it says 401 Unauthorized. It should be set to None in config/session.php

Please, see links below: 
https://blog.chromium.org/2020/02/samesite-cookie-changes-in-february.html
https://github.com/laravel/airlock/issues/11#issuecomment-596438725