Ignore APP_URL when it's not set (#279)
If APP_URL is missing then null is passed to parse_url(), which breaks
on PHP 8.1 even when you're only running composer install. On PHP <==8
it wouldn't error, but you'd have an empty entry in sanctum.stateful.

Setting SANCTUM_STATEFUL_DOMAINS in the project's .env isn't enough
because the default config is still merged in the service provider as
long as config isn't cached (which it won't be during composer install).

---

Also removes APP_URL from phpunit.xml, setting in the test that needs it
instead. By using getEnvironmentSetUp instead of @environment-setup, we
don't need putenv everywhere and it more closely matches other tests.