Carry through session env settings
We have unique use cases where we do not want http_only and same_site hardcoded here. We need them to be specific values that we have set in our ENV. 

This change will allow it to pull the values from the ENV if they exist, if not, fallback to the original values. Please consider, thanks!