Fix return type using Intersection Types (PHP 8.1)
The `currentAccessToken()` method was updated to fix the return type declaration. Intersection Types from PHP 8.1 were used to specify that the returned object is an instance of `Illuminate\Database\Eloquent\Model` and also implements the `Laravel\Sanctum\Contracts\HasAbilities` interface.

By utilizing the `Model&HasAbilities` type in the PHPDoc, we accurately document that the method returns an object that satisfies both the `Model` class requirement and the `HasAbilities` interface implementation.

Note: Please ensure that your Laravel project is updated to the latest version, as it requires PHP 8.1 or a later version to utilize Intersection Types.